Started abstraction of render interface to be able to implement Vulkan and OpenGL as options for the time being
