+++ examples/Basic/materials.go
@@ -8,7 +8,6 @@ import (
 	"Gopher3D/internal/renderer"
 
 	mgl "github.com/go-gl/mathgl/mgl32"
-	// Example Perlin noise library
 )
 
 type GoCraftBehaviour struct {
@@ -27,6 +26,7 @@ func main() {
 	NewMaterialExampleBehaviour(engine)
 	engine.Render(768, 50)
 }
+
 func (mb *GoCraftBehaviour) Start() {
 	createWorld(mb)
 }
@@ -38,11 +38,11 @@ func (mb *GoCraftBehaviour) Update() {
 func createWorld(mb *GoCraftBehaviour) {
 	model, _ := loader.LoadObjectWithPath("../resources/obj/f104starfighter.obj", true)
 	mb.SceneModel = model
-	spawnBlock(mb.SceneModel, 0, 0)
+	spawnBlock(mb.engine, mb.SceneModel, 0, 0)
 }
 
-func spawnBlock(model *renderer.Model, x, z int) {
+func spawnBlock(engine *engine.Gopher, model *renderer.Model, x, z int) {
 	model.SetPosition(0, 0, 0)
 	model.Scale = mgl.Vec3{20.0, 20.0, 20.0}
-	renderer.AddModel(model)
+	engine.AddModel(model)
 }
